+++ R/search.r
@@ -136,7 +136,7 @@ pretty.filesize <- function(x) {
 #' ESGF search functionality (and can work with aggregations and files as well).
 #'
 #' @param ... A list of constraints on which to apply the search.
-#' @param node URL (including "http://") of the ESGF index node to search
+#' @param index.node URL (including "http://") of the ESGF index node to search. Defaults to options("resgf.indexNode")
 #' @param search.limit Maximum number of values to return in the search. ESGF currently limits this to 10000
 #' @param show.all.replicas Show all copies (replicas) of a file or dataset (TRUE) or just the "original". Defaults to FALSE
 #' @param show.all.versions Should the result showsall versions of a file or dataset? Defaults to FALSE
@@ -155,7 +155,7 @@ pretty.filesize <- function(x) {
 #'                       project="CMIP6")
 resgf_search <-
   function(...,
-           node="http://esgf-node.llnl.gov/esg-search",
+           index.node=getOption("resgf.indexNode"),
            search.limit=10000,
            show.all.replicas=FALSE,
            show.all.versions=FALSE,
@@ -176,7 +176,7 @@ resgf_search <-
     
     #Build search command
     search.cmd <- sprintf("%s/search?format=application%%2Fsolr%%2Bjson&limit=%i&%s",
-                          node,
+                          index.node,
                           search.limit,
                           facet.constraints)
     if(!show.all.replicas) search.cmd <- paste0(search.cmd,'&replica="false"')
@@ -283,11 +283,17 @@ resgf_simplify <- function(object) {
 #' @param max.files maximum number of files to be returned in a chunk
 #' @param max.datasets maximum number of datasets to request in a chunk
 #' @param processes Number of processes to perform in parallel
+#' @param index.node URL (including "http://") of the ESGF index node to search. Defaults to options("resgf.indexNode"). 
+#' Passed in this instance further to `resgf_search*()`
 #'
 #' @return
 #' @export
 resgf_get_filelist <-
-  function(object,max.files=1000,max.datasets=10,processes=1) {
+  function(object,
+           max.files=1000,
+           max.datasets=10,
+           processes=1,
+           index.node=getOption("resgf.indexNode")) {
     #Require input object to be dataset search result
     assert_that(max(object$number_of_files)< max.files,
                 msg="File chunk size is too small.")
@@ -318,9 +324,10 @@ resgf_get_filelist <-
     this.fileset <-
       pblapply(chunk.l,
                function(x) {resgf_search_files(dataset_id=x$id,
-                                               simplify=FALSE)},
+                                               simplify=FALSE,
+                                               index.node=index.node)},
                cl=processes)
-    
+
     #And we're done (with some tidying)
     rtn <-
       this.fileset %>%